% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_similarity.R
\name{ang_sim_check}
\alias{ang_sim_check}
\title{check inputs for angular_angular similarity}
\usage{
ang_sim_check(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\description{
helper function
}

