% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vector.R
\name{average_vector}
\alias{average_vector}
\title{Calculates and average vector}
\usage{
average_vector(x, fun = "mean", ...)
}
\arguments{
\item{x}{Matrix of numerical data. Columns are components of the data, row per vector}

\item{fun}{function to use, current options: mean, median}

\item{...}{additional arguments to be passed to fun}
}
\value{
Mean vector
}
\description{
Calculates an average vector from rows of multiple vectors
}
\examples{
x <- matrix(c(1,5,2,10,3,10), ncol = 2)
average_vector(x)
average_vector(x, fun = "median")
average_vector(x, fun = "median", na.rm = TRUE)
}

