% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vector.R
\name{colMedians}
\alias{colMedians}
\title{column median}
\usage{
colMedians(x, ...)
}
\arguments{
\item{x}{dataframe of matrix}

\item{...}{additional arguments to be passed to median}
}
\description{
interal function
}

