% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_pairs.R
\name{cosine_pairs}
\alias{cosine_pairs}
\title{Calculate pairs of cosine similarities between replicates}
\usage{
cosine_pairs(x, cols)
}
\arguments{
\item{x}{list, each element being a separate compound}

\item{cols}{integer, column indices that match numeric data (e.g principal
comonents)}
}
\value{
dataframe of compound combinations across replicates with a column
   of cosine similarity values
}
\description{
Given a list, with an element per compound, \code{cols}
are the elements of the vector with which to calculate the cosine
similarity. \code{cosine_pairs} will calculate the theta value between all
replicates for each possible pairing of compounds and return the result
in a long-format dataframe.
}
\examples{
cmpds <- c(rep('compound_1', 100),
           rep('compound_2', 100),
           rep('compound_3', 100))
replicate <- rep(1:100, 3)
PC1 <- rnorm(300)
PC2 <- rnorm(300)
df <- data.frame(cmpds, replicate, PC1, PC2)
df_split <- split(df, df$cmpds)
# works with unequal replicate sizes
df_split$a <- df_split$a[-c(1:10), ]
cosine_pairs(df_split, 3:4)
}

