% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sim.R
\name{cosine_sim_vec_check_input}
\alias{cosine_sim_vec_check_input}
\title{Check inputs for cosine_sim_vector}
\usage{
cosine_sim_vec_check_input(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\description{
helper function for \code{cosine_sim_vector}
}

